refactor : 리뷰 내용 수정

+++ monew-api/src/main/java/com/monew/monew_api/interest/dto/response/InterestDto.java
@@ -6,7 +6,7 @@ public record InterestDto(
     Long id,
     String name,
     List<String> keywords,
-    int subscriberCount,
+    Long subscriberCount,
     boolean subscribedByMe
 ) {
